+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.17.2] - 2024-03-27
+### Added
+- Add code to the beginning of some CFS functions to check if they have been called without
+  necessary arguments, and if so, to log a warning and return immediately.
+- Added similar code to some PCS functions.
+- Created `PowerControlComponentsEmptyException`; raise it when some PCS functions receive
+  empty component list arguments.
+
+### Changed
+- If the status operator `_run` method finds no enabled components, stop immediately, as there is
+  nothing to do.
+
 ## [2.17.1] - 2024-03-21
 ### Changed
 - Improvements to BOS v2 debug logging.